added multiple filter components
